requirements.txt
langchain
langchain-core
langchain-openai
openai
langchain
python-dotenv
langchain_huggingface
langchain_community
langchain_google_genai
sentence-transformers
streamlit
